cdk-app/pyproject.toml
[tool.poetry]
name = "envs"
version = "1.0.0"
description = ""
authors = ["Your Name <tomer.shoob@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "2.94.0"
constructs = "^10.3.0"
boto3 = "^1.34.93"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
